+++ bingads/v12/bulk/entities/__init__.py
@@ -26,3 +26,4 @@ from .audiences import *
 from .target_criterions import *
 from .labels import *
 from .bulk_offline_conversion import *
+from .bulk_experiment import *