+++ bingads/v12/bulk/entities/ad_extensions/__init__.py
@@ -9,3 +9,4 @@ from .bulk_review_ad_extensions import *
 from .bulk_structured_snippet_ad_extensions import *
 from .bulk_sitelink_ad_extensions import *
 from .bulk_price_ad_extensions import *
+from .bulk_action_ad_extensions import *